support create column as expr
