+++ src/expr.js
@@ -1,9 +1,10 @@
-import { literalToSQL, columnRefToSQL, toUpper, connector } from './util'
+import { literalToSQL, toUpper, connector } from './util'
 import { alterExprToSQL } from './alter'
 import { aggrToSQL } from './aggregation'
 import { assignToSQL } from './assign'
 import { binaryToSQL } from './binary'
 import { caseToSQL } from './case'
+import { columnRefToSQL } from './column'
 import { castToSQL, funcToSQL } from './func'
 import { intervalToSQL } from './interval'
 import { selectToSQL } from './select'
